+++ src/PJ/fli/isempty.c
@@ -3,5 +3,5 @@
 Boolean pj_i_is_empty_rec(Fli_frame *frame)
 /* returns TRUE if the *frame is an empty delta record */
 {
-	return(frame->size <= sizeof(Fli_frame));
+	return(frame->size <= (long)sizeof(Fli_frame));
 }
